+++ examples/lean_chacha20/src/lib.rs
@@ -50,7 +50,7 @@ theorem Lean_chacha20.chacha20_quarter_round_spec a b c d state:
   intros
   mvcgen [Lean_chacha20.chacha20_quarter_round,
           Lean_chacha20.chacha20_line,
-          Result.ofOption, ]
+          RustM.ofOption, ]
   <;> try omega
 "
 )]