+++ hax-lib/core-models/src/core/option.rs
@@ -4,6 +4,8 @@ pub enum Option<T> {
 }
 
 #[hax_lib::fstar::before("open Rust_primitives.Integers")]
+struct Dummy;
+
 use super::default::Default;
 use super::ops::function::*;
 use super::result::Result::*;