+++ tests/lean-tests/src/lib.rs
@@ -4,12 +4,15 @@
 pub mod comments;
 pub mod constants;
 pub mod enums;
+pub mod floats;
 pub mod ite;
 pub mod loops;
+pub mod matching;
 pub mod monadic;
 pub mod reject_do_dsl;
 pub mod structs;
 pub mod traits;
+pub mod types;
 
 const FORTYTWO: usize = 42;
 const MINUS_FORTYTWO: isize = -42;
@@ -61,5 +64,6 @@ fn binop_resugarings(x: u32) -> u32 {
     let rem = mul % 4;
     let div = rem / 5;
     let rshift = div >> x;
+    let lshift = div << x;
     x
 }