+++ tests/lean-tests/src/matching.rs
@@ -1,0 +1,26 @@
+fn test_const_matching(x: u32, c: char, s: &str, b: bool) -> u32 {
+    let x = match x {
+        0 => 42,
+        _ => 0,
+    };
+    let c = match c {
+        'a' => 42,
+        _ => 0,
+    };
+    let s = match s {
+        "Hello" => 42,
+        _ => 0,
+    };
+    let b = match b {
+        true => 42,
+        false => 0,
+    };
+    return x + c + s + b;
+}
+
+fn test_binding_subpattern_matching(x: (u8, (u8, u8))) -> u8 {
+    match x {
+        (0, pair @ (a, b)) => a + b + pair.0 + pair.1,
+        _ => 0,
+    }
+}